feat: メソッド分割
Closes #1
